requirements.txt
natsort==5.1.1
pydicom==0.9.9
PyYAML==3.12
